% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{email_interface}
\alias{email_interface}
\title{Email interface}
\usage{
email_interface(email_args_list, viewer = "dialog", height = 800, width = 800)
}
\arguments{
\item{email_args_list}{A list as created by `email_template` or a list of those lists.}

\item{viewer}{One of dialog, browser or pane. Where the shinyapp will be executed.}

\item{height}{The height when viewer is set to dialog.}

\item{width}{The width when viewer is set to dialog.}
}
\description{
Email interface
}
\details{
A simple shiny app that lets you see the details of the emails you composed. You can
browser the emails and send one or more. A sent email can't be sent again. If you sent one or more
and then click in 'Send all', it will send only the ones that were not sent yet.
}
