% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{email_template}
\alias{email_template}
\title{An email template}
\usage{
email_template(
  from = NULL,
  to = NULL,
  cc = NULL,
  bcc = NULL,
  subject = NULL,
  title = NULL,
  body = NULL,
  header = NULL,
  footer = NULL,
  attachments = NULL
)
}
\arguments{
\item{from}{Your email}

\item{to}{A vector of emails}

\item{cc}{A vector of emails}

\item{bcc}{A vector of emails}

\item{subject}{A string with the subject}

\item{title}{The title of the email (see `blastula`)}

\item{body}{Where the main text goes}

\item{header}{A string with the header text}

\item{footer}{A string with the footer text}

\item{attachments}{A vector with attachments names (local paths or absolute paths)}
}
\value{
A named list used to create the email
}
\description{
An email template
}
