% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_email_credentials}
\alias{get_email_credentials}
\title{Internal function to create the credentials file used in blastula}
\usage{
get_email_credentials(email)
}
\arguments{
\item{email}{The email adress.}
}
\value{
A credentials file
}
\description{
Internal function to create the credentials file used in blastula
}
