% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.error}
\alias{is.error}
\title{Check if an object has class try-error}
\usage{
is.error(x)
}
\arguments{
\item{x}{Any object#'}
}
\value{
FALSE or TRUE
}
\description{
Check if an object has class try-error
}
