% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{limpa_string}
\alias{limpa_string}
\title{Clean the email string to save it as a valid keyring name}
\usage{
limpa_string(x)
}
\arguments{
\item{x}{a string}
}
\value{
string without . or - or \@
}
\description{
Clean the email string to save it as a valid keyring name
}
