% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_email}
\alias{prepare_email}
\title{Preparing the email}
\usage{
prepare_email(email_args)
}
\arguments{
\item{emails_args}{A list as created by `email_template`}
}
\value{
An object suited to be passed to `blastula`
}
\description{
Preparing the email
}
