% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{register_email_address}
\alias{register_email_address}
\title{Register the email in a keyring for future use}
\usage{
register_email_address(email)
}
\arguments{
\item{email}{The email adress.}
}
\description{
Register the email in a keyring for future use
}
\details{
This function creates a keyring with your email and additional parameters (password, email provider, etc)
used in the `credentials` parameters of blastula `smtp_send`. Some dialogs will pop asking for your email
password, a PIN code used to retrieve this password and so on.
}
