% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{send_email}
\alias{send_email}
\title{Sends the emails}
\usage{
send_email(email_args_list, credentials = NULL, .function = NULL, ...)
}
\arguments{
\item{email_args_list}{A list as created by `email_template` or a list of those lists.}

\item{credentials}{Used to authenticate blastula. If NULL, it gets a credential object from `get_email_credentials`.}
}
\description{
Sends the emails
}
